pet_connect_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter>=2.4.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
Pillow==10.1.0
numpy
scikit-learn
pandas
django-cors-headers